requirements.txt
Flask==3.1.0 
mysql-connector-python==9.3.0 
pandas==2.2.3 
PyPDF2==3.0.1 
requests==2.32.3 
python-dotenv==1.1.0 
gunicorn==23.0.0
bcrypt==4.3.0
Flask-Login==0.6.3
twilio==9.5.2